Deleted a lot of the old Rune types
